+++ maestro/src/__init__.py
@@ -17,8 +17,8 @@ import dotenv
 dotenv.load_dotenv()
 
 from .workflow import Workflow
-from .bee_agent import BeeAgent
-from .agent import save_agent, restore_agent, remove_agent
+from .agents.bee_agent import BeeAgent
+from .agents.agent import save_agent, restore_agent, remove_agent
 from .deploy import Deploy
 
 __all__ = {
